longvector: support bitcast of scalars/vectors
regroup work on bitcast for replacepointerbitcast into BitcastUtils in
order to use it in LongVectorLowering
